--- components/context/todoContext.tsx
@@ -1,13 +1,0 @@
-'use client'
-
-import { FC, createContext } from "react"
-
-export const TodoContext = createContext('')
-
-export const TodoContextProvider: FC = (props: any) => { 
-  return (
-    <TodoContext.Provider value={props.value}>
-      {props.children}
-    </TodoContext.Provider>
-  )
-}
